+++ ec/__init__.py
@@ -1,4 +1,4 @@
 
 __author__ = 'Laufire Technologies'
 __email__ = 'laufiretechnologies@gmail.com'
-__version__ = '0.2.1'
+__version__ = '0.2.2'