+++ ec/modules/exposed.py
@@ -5,34 +5,31 @@ exposed
 A module for members that are used by ec, as well as exposed to the scripts.
 
 """
-def get(desc, type=None, autoDesc=True, **Kwargs):
+def get(desc='', type=None, **KwArgs):
   """Helps to interactively get user input.
   
   Args:
     desc (str): The description for input.
     type (type / CustomType): The type of the input (defaults to None).
-    autoDesc (bool): When set to false, it suppresses the generation of the prompt string based on the given arguments (defaults to True).
-    default: The default value for the input.
     
   Notes:
-    * The arg **default** is made to be a part pf Kwargs, rather than an explicit argument, in order to allow the declaration of 'None' as the default value.
+    * The arg **default** is made to be a part pf KwArgs, rather than an explicit argument, in order to allow the declaration of 'None' as the default value.
+    * KwArgs acts as a data container for unexpected attibutes that are used by underlying helpers.
   """
-  has_default = 'default' in Kwargs
-  default = Kwargs.get('default')
-  label = '%s: ' % ('%s%s' % (desc, ', %s' % type if isinstance(type, CustomType) else '') if autoDesc else desc)
+  if not desc:
+    desc = '%s%s: ' % (getAutoDesc(KwArgs), ', %s' % type if isinstance(type, CustomType) else '')
   
   while True:
     try:
-      got = raw_input(label)
+      got = raw_input(desc)
       
     except EOFError:
+      got = None
+      
+    if not got and 'default' in KwArgs:
+      return KwArgs['default']
     
-      got = ''
-      
     try:
-      if has_default and not got:
-        return default
-      
       return type(got) if type else got
       
     except ValueError:
@@ -56,5 +53,5 @@ def static(cls):
   return cls
   
 # Cross dependencies
-from helpers import err
+from helpers import err, getAutoDesc
 from classes import CustomType