+++ ec/types/regex.py
@@ -11,9 +11,13 @@ from ..modules.classes import CustomType
 class pattern(CustomType):
   """Get inputs that fit a specific pattern.
   """
-  def __init__(self, pattern, flags=0, desc=None):
-    CustomType.__init__(self, desc)
+  def __init__(self, pattern, flags=0, **Config):
     self.exp = re.compile(pattern, flags)
+    
+    if not 'type_str' in Config:
+      Config['type_str'] = 'a string matching the pattern \'%s\'' % self.exp.pattern
+      
+    CustomType.__init__(self, **Config)
     
   def __call__(self, val):
     if not self.exp.match(val):
@@ -21,11 +25,9 @@ class pattern(CustomType):
       
     return val
     
-  def __str__(self):
-    return getattr(self, 'desc', 'a string matching the pattern \'%s\'' % self.exp.pattern)
-    
-email = pattern('^([a-z0-9_\.-]+)@([\da-z\.-]+)\.([a-z\.]{2,6})$', desc='Email')
-username = pattern('^[a-z0-9_-]{3,16}$', desc='Username')
-password = pattern('^[A-z0-9_-]{6,18}$', desc='Password')
-slug = pattern('^[a-z0-9-]+$', desc='Slug')
-number = pattern('^[0-9]+$', desc='Number') # a positive integer
+# Quick Templates
+email = pattern('^([a-z0-9_\.-]+)@([\da-z\.-]+)\.([a-z\.]{2,6})$', type_str='Email')
+username = pattern('^[a-z0-9_-]{3,16}$', type_str='Username')
+password = pattern('^[A-z0-9_-]{6,18}$', type_str='Password')
+slug = pattern('^[a-z0-9-]+$', type_str='Slug')
+number = pattern('^[0-9]+$', type_str='Number') # a positive integer