+++ ec/utils.py
@@ -9,17 +9,16 @@ from modules.classes import CustomType
 
 __all__ = ['get', 'static', 'custom']
   
-def custom(validator, desc=None, converter=None):
+def custom(validator, converter=None, **Config):
   """Helps with creating dynamic CustomTypes on the fly.
   
   Args:
     validator (callable): that acts as the converter / validator.
-    desc (str): the desc to be dispalyed on the helps. Defaults to None.
     converter (callable):  A callable that converts the given value. Defaults to None.
   """
   class dynamic(CustomType):
     def __init__(self):
-      CustomType.__init__(self, desc)
+      CustomType.__init__(self, **Config)
       
     def __call__(self, val):
       if converter: